# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Budget Pulse Watch API
  version: 1.0.0
  description: |
    REST API for PRF Monitoring Dashboard. Includes authentication, PRF import, budgets, COA, and settings.
    Protected endpoints accept either:
      - Bearer JWT in the Authorization header, or
      - API Key via header `x-api-key: <token>`, or
      - API Key via header `Authorization: ApiKey <token>`
servers:
  - url: /
    description: Same origin
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Provide API key directly in header: `x-api-key: <token>`
    ApiKeyAuthorization:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Provide API key using custom Authorization scheme: `Authorization: ApiKey <token>`
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, doccon, user]
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: string
        authType:
          type: string
          enum: [local, ldap]
        createdAt:
          type: string
          format: date-time
    AuthLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Username or email
        password:
          type: string
          format: password
        authType:
          type: string
          enum: [auto, local, ldap]
          default: auto
    AuthLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
paths:
...
  /api/auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          description: Missing parameters
  /api/auth/api-keys:
    post:
      summary: Create API key
      description: Create a new API key. Returns the plaintext token once.
      tags: [API Keys]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name:
                  type: string
                role:
                  type: string
                  enum: [admin, doccon, user]
                expiresAt:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Unauthorized
    get:
      summary: List API keys
      tags: [API Keys]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Keys
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Unauthorized
  /api/auth/api-keys/{id}:
    delete:
      summary: Deactivate API key
      tags: [API Keys]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deactivated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Unauthorized
        '400':
          description: Invalid id
  /api/auth/verify:
    get:
      summary: Verify JWT token
      tags: [Auth]
      security:
        - bearerAuth: []
        - ApiKeyHeader: []
        - ApiKeyAuthorization: []
      responses:
        '200':
          description: Valid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /api/auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      security:
        - bearerAuth: []
        - ApiKeyHeader: []
        - ApiKeyAuthorization: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
...
  /api/auth/logout:
    post:
      summary: Logout
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
...